chore move fast test where it belongs
