+++ node/subsystem/src/lib.rs
@@ -221,3 +221,64 @@ impl<C: SubsystemContext> Subsystem<C> for DummySubsystem {
 		}
 	}
 }
+
+
+
+
+#[cfg(test)]
+mod tests {
+	use super::*;
+	use std::time::Duration;
+	use futures::{executor, future, Future};
+	use assert_matches::assert_matches;
+
+	macro_rules! assert_msg {
+		($msg:expr, $expected:expr) => {
+			{
+				assert_matches!(
+					$msg,
+					AllMessages::TestVal(x)
+					=> {
+						assert_eq!(x, $expected);
+					}
+				)
+			}
+		};
+	}
+
+	macro_rules! msg {
+		($val:expr) => {
+			FromOverseer::Communication {
+				msg: AllMessages::TestVal($val)
+			}
+		};
+	}
+
+	#[test]
+	fn fast() {
+		let pool = sp_core::testing::TaskExecutor::new();
+		let (mut ctx, mut overseer) = make_subsystem_context(pool);
+
+		let ping = async move {
+			for x in 50_000usize..100_000 {
+				ctx.send_message(AllMessages::TestVal(x)).await.unwrap();
+				let _y = ctx.recv().await.unwrap();
+			}
+		};
+
+		let pong = async move {
+
+			for y in 0usize..50_000 {
+				let x = overseer.recv().await;
+				assert_msg!(x, y+50_000);
+				overseer.send(msg!(y) ).await;
+
+			}
+		};
+
+		futures::pin_mut!(ping);
+		futures::pin_mut!(pong);
+
+		executor::block_on(future::join(ping, pong));
+	}
+}
